+++ src/Compute/Compute.Test/ScenarioTests/GalleryTests.cs
@@ -93,5 +93,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
             TestRunner.RunTestScript("Test-InVMAccessControlProfileVersion");
         }
 
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestGalleryManagedIdentity()
+        {
+            TestRunner.RunTestScript("Test-GalleryManagedIdentity");
+        }
+
     }
 }